+++ app/Http/Controllers/CategoriesController.php
@@ -7,6 +7,11 @@ use App\Category;
 
 class CategoriesController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware("auth");
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,6 +19,11 @@ class CategoriesController extends Controller
      */
     public function index()
     {
+        if(auth()->user()->role != "admin")
+        {
+            return redirect("/")->with("error", "Unauthorized");
+        }
+
         $categories = Category::all();
         return view("categories/index")->with("categories", $categories);
     }
@@ -36,6 +46,12 @@ class CategoriesController extends Controller
      */
     public function store(Request $request)
     {
+
+        if(auth()->user()->role != "admin")
+        {
+            return redirect("/")->with("error", "Unauthorized");
+        }
+
         $this->validate($request,
         [
             "name" => "required"
@@ -66,6 +82,11 @@ class CategoriesController extends Controller
      */
     public function edit($id)
     {
+        if(auth()->user()->role != "admin")
+        {
+            return redirect("/")->with("error", "Unauthorized");
+        }
+
         $category = Category::find($id);
         return view("categories/edit")->with("category", $category);
     }
@@ -79,6 +100,11 @@ class CategoriesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(auth()->user()->role != "admin")
+        {
+            return redirect("/")->with("error", "Unauthorized");
+        }
+
         $this->validate($request,
         [
             "name" => "required"
@@ -98,6 +124,11 @@ class CategoriesController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->role != "admin")
+        {
+            return redirect("/")->with("error", "Unauthorized");
+        }
+
         $category = Category::find($id);
         $category->delete();
         return redirect("/categories")->with("success", "Category deleted!");
